refactor: move interfaces to own folder
