+++ src/interfaces/access-token.ts
@@ -1,0 +1,5 @@
+export interface AccessToken {
+	token: string
+	type: string
+	expiresAt: number
+}
