+++ src/interfaces/airthings-configuration.ts
@@ -1,0 +1,4 @@
+export interface AirThingsConfiguration {
+	id: string
+	secret: string
+}
