+++ src/interfaces/device.ts
@@ -1,0 +1,15 @@
+export interface Device {
+	id: string
+	deviceType: string
+	sensors: string[]
+	segment: {
+		id: string
+		name: string
+		started: string
+		active: boolean
+	}
+	location: {
+		id: string
+		name: string
+	}
+}
