+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export type { AccessToken } from './access-token'
+export type { AirThingsConfiguration } from './airthings-configuration'
+export type { Device } from './device'
+export type { Readings } from './readings'
