+++ src/interfaces/readings.ts
@@ -1,0 +1,43 @@
+export interface Readings {
+	battery: number
+	co2: number
+	humidity: number
+	light: number
+	lux: number
+	mold: number
+	pm1: number
+	pm10: number
+	pm25: number
+	pressure: number
+	pressureDifference: number
+	radonShortTermAvg: number
+	rssi: number
+	sla: number
+	temp: number
+	time: number
+	virusRisk: number
+	voc: number
+	outdoorTemp: number
+	outdoorHumidity: number
+	outdoorPressure: number
+	outdoorPm1: number
+	outdoorPm10: number
+	outdoorPm25: number
+	outdoorNo2: number
+	outdoorO3: number
+	outdoorSo2: number
+	outdoorCo: number
+	outdoorNo: number
+	controlSignal: number
+	controlSignal01: number
+	controlSignal02: number
+	controlSignal03: number
+	controlSignal04: number
+	controlSignal05: number
+	controlSignal06: number
+	controlSignal07: number
+	controlSignal08: number
+	regulationPressure: number
+	regulationHeigh: number
+	relayDeviceType: string
+}
